+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ValueAssert.java
@@ -553,8 +553,8 @@ public class ValueAssert {
     }
 
     public interface NonFunctionalInterface {
-        void foobarbaz();
-
+        default void foobarbaz() {
+        }
     }
 
     @FunctionalInterface